Added relationship between Link and Page

+++ database/migrations/2020_04_06_085417_create_pages_table.php
@@ -18,7 +18,7 @@ class CreatePagesTable extends Migration
             $table->bigInteger('assignment_id')->unsigned();
             $table->string('name');
             $table->string('description');
-            $table->integer('address');
+            $table->bigInteger('address');
             $table->timestamps();
             $table->foreign('assignment_id')->references('id')->on('assignments')
